[CIT-500] Move all the messages into the `messages` module
